Be explicit about likely need to "re-join" a group in call for review
Per request from a member during the 2019 Second Screen WG charter review. 
Via François Daoust: The members run lengthy legal processes before they can join a group and typically need to re-run that process when that happens. They could start it before the Call for Participation is issued but for that to happen, they need to know beforehand that this will actually be needed.